data results officially redirect onto a new page
