+++ client/package.json
@@ -7,6 +7,8 @@
     "google-maps-react": "^2.0.2",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4"
   },
   "scripts": {
